Revert "Revert "Introduce state-specific employee accounts""
This reverts commit 75d3dfca205f8636fdd38ee54551a472d4d2c0b4.

Reintroduce state-specific accounts for fiduciary liabilities. After
some testing, I discovered that analytic accounts were not a good fit
for this because they are more intended for nominal accounts, not for
balance sheet liabilities, and designed for all analytic axes to be used
together, not in a cherry-picking sort of way.